remove legacy parts update readme and add test manifest
