--- docs/tests.md
@@ -1,127 +1,0 @@
-
-## cdperf Tests
-
-This lists and briefly explains the tests that are done in the K6 based test framework cdperf.
-
-### Most used file sizes upload
-
-Script: `/root/cdperf/tests/k6/test-issue-github-enterprise-4115-most-used-sizes-upload.js`
-
-This test creates a pattern of the so called *most used file sizes*. It contains a collection of test files in sizes how we find them in typical larger ownCloud installations. Specifically, that contains 84% files with a size smaller than 1MB, 12.5% files between 1MB and 10MB, 3% files between 1MB and 100MB and 0.8% files larger than 100MB.
-
-The files are created, uploaded, downloaded and afterwards deleted in this test. The current implementation tests within a total of 15GB data per User.
-
-This simulates a "normal" upload/download scenario with a common file set.
-
-### Deep Propfind
-
-Scripts:
-* `/root/cdperf/tests/k6/test-issue-github-ocis-1018-propfind-deep-100-files-45-nested-folders.js`
-* `/root/cdperf/tests/k6/test-issue-github-ocis-1018-propfind-deep-1000-files-5-nested-folders.js`
-
-This test creates a nested directory structure and puts some files into the folder.
-
-The depth and file count depends on the test that is running.
-
-After provisioning is complete, the test does one propfind per folder
-
-This simulates the ETag check of a large nested directory.
-
-### Flat Propfind
-
-Script: `/root/cdperf/tests/k6/test-issue-github-ocis-1018-propfind-flat-1000-files.js`
-
-This test creates many (1000 a 1kB) files in the root of the cloud.
-
-The files are uploaded and afterwards, a PROPFIND to the root directory is done.
-
-This test gives an indication on the PROPFIND speed for a larger structure of nested folders and files and also measures the pure upload speed of small files.
-
-### Upload Big Files
-
-Script: `/root/cdperf/tests/k6/test-issue-github-ocis-1018-upload-download-delete-many-large.js`
-
-In total this test creates 5 GB of files which are distributed from 5 MB up to 1 GB per file.
-
-The files are uploaded, downloaded and deleted afterwards.
-
-This gives an indication about the transmission speed of big files.
-
-### Upload Small Files
-
-Script: `/root/cdperf/tests/k6/test-issue-github-ocis-1018-upload-download-delete-many-small.js`
-
-This test creates files between 500 kB and 25 MB, with a total size of 1.7 GB.
-
-The files are uploaded, downloaded and deleted.
-
-This measures the up- and download speed of smaller files.
-
-### Upload, Delete and Trash Big Files
-
-Script: `/root/cdperf/tests/k6/test-issue-github-ocis-1018-upload-delete-trash-many-large.js`
-
-In total this test creates 5 GB of files which are distributed from 5 MB up to 1 GB per file.
-
-The files are uploaded, deleted and removed from trash afterwards.
-
-This gives an indication about the transmission speed of big files and tests the performance of trashing those files.
-
-### Upload, Delete and Trash Small Files
-
-Script: `/root/cdperf/tests/k6/test-issue-github-ocis-1018-upload-delete-trash-many-small.js`
-
-This test creates files between 500 kB and 25 MB, with a total size of 1.7 GB.
-
-The files are uploaded, deleted and removed from trash afterwards.
-
-This gives an indication about the transmission speed of smaller files and tests the performance of trashing those files.
-
-### Upload, Delete and Restore Big Files
-
-Script: `/root/cdperf/tests/k6/test-issue-github-ocis-1018-upload-delete-restore-many-large.js`
-
-In total this test creates 5 GB of files which are distributed from 5 MB up to 1 GB per file.
-
-The files are uploaded, deleted and restored from trash afterwards.
-
-This gives an indication about the transmission speed of big files and tests the performance of restoring those files from trash.
-
-### Upload, Delete and Restore Small Files
-
-Script: `/root/cdperf/tests/k6/test-issue-github-ocis-1018-upload-delete-restore-many-small.js`
-
-This test creates files between 500 kB and 25 MB, with a total size of 1.7 GB.
-
-The files are uploaded, deleted and restored from trash afterwards.
-
-This gives an indication about the transmission speed of smallert files and tests the performance of restoring those files from trash.
-
-### Upload and Download with New User
-
-Script: `/root/cdperf/tests/k6/test-issue-github-ocis-1018-upload-download-delete-with-new-user.js`
-
-This script creates a small amount of 1 kB files.
-
-The files are uploaded, downloaded and deleted with a freshly created user.
-
-### Propfind and Rename
-
-Script: `/root/cdperf/tests/k6/test-issue-github-ocis-1399-propfind-deep-rename.js`
-
-The script creates 50 folders in the root with in sum 10 MB content files.
-
-The folders and files are uploaded, renamed and after the rename, a PROPFIND is done.
-
-This reflects the behaviour of renames of folders.
-
-### Share with User
-
-Script: `/root/cdperf/tests/k6/test-issue-github-ocis-1399-share-with-new-user.js`
-
-This script creates a folder and provisions the folder with some files right after.
-
-Right after the test folder is uploaded, it will be shared with a different user. With the share receiver, the files are downloaded. Finally, the share is removed and the user and files are deleted.
-
-This reflects the transmission performance of shared folders.
-
