+++ package.json
@@ -34,7 +34,7 @@
         "@rollup/plugin-json": "^4.0.3",
         "@rollup/plugin-node-resolve": "^7.1.3",
         "@rollup/pluginutils": "^4.1.0",
-        "@types/k6": "^0.41.0",
+        "@types/k6": "^0.42.0",
         "@types/lodash": "^4.14.165",
         "@types/xmldom": "^0.1.30",
         "@typescript-eslint/eslint-plugin": "^4.9.0",
