+++ run/share-upload-rename-default.js
@@ -1,0 +1,1 @@
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var e=require("k6/crypto"),t=require("k6/execution"),n=require("k6"),r=require("k6/http"),o=require("k6/encoding");function a(e){return e&&"object"==typeof e&&"default"in e?e:{default:e}}var i=a(t),u=a(r),s=a(o);function c(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function l(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function f(e,t,n){return t&&l(e.prototype,t),n&&l(e,n),e}function p(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}function d(e,t){var n=Object.keys(e);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(e);t&&(r=r.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),n.push.apply(n,r)}return n}function h(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{};t%2?d(Object(n),!0).forEach((function(t){p(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):d(Object(n)).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}function v(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&g(e,t)}function m(e){return(m=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function g(e,t){return(g=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function b(e,t){return!t||"object"!=typeof t&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function y(e){var t=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(e){return!1}}();return function(){var n,r=m(e);if(t){var o=m(this).constructor;n=Reflect.construct(r,arguments,o)}else n=r.apply(this,arguments);return b(this,n)}}function w(e,t){var n=t.get(e);if(!n)throw new TypeError("attempted to get private field on non-instance");return n.get?n.get.call(e):n.value}function N(e,t,n){var r=t.get(e);if(!r)throw new TypeError("attempted to set private field on non-instance");if(r.set)r.set.call(e,n);else{if(!r.writable)throw new TypeError("attempted to set read only private field");r.value=n}return n}var _=function(e,t){for(var n=-1,r=Array(e);++n<e;)r[n]=t(n);return r};var E=function(e){return e};var T=function(e){return"function"==typeof e?e:E};var O=function(e){var t=typeof e;return null!=e&&("object"==t||"function"==t)},x="undefined"!=typeof globalThis?globalThis:"undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof self?self:{};function S(e,t){return e(t={exports:{}},t.exports),t.exports}var D="object"==typeof x&&x&&x.Object===Object&&x,A="object"==typeof self&&self&&self.Object===Object&&self,I=D||A||Function("return this")(),R=I.Symbol,C=Object.prototype,k=C.hasOwnProperty,j=C.toString,U=R?R.toStringTag:void 0;var P=function(e){var t=k.call(e,U),n=e[U];try{e[U]=void 0;var r=!0}catch(e){}var o=j.call(e);return r&&(t?e[U]=n:delete e[U]),o},L=Object.prototype.toString;var M=function(e){return L.call(e)},q=R?R.toStringTag:void 0;var F=function(e){return null==e?void 0===e?"[object Undefined]":"[object Null]":q&&q in Object(e)?P(e):M(e)};var z=function(e){return null!=e&&"object"==typeof e};var $=function(e){return"symbol"==typeof e||z(e)&&"[object Symbol]"==F(e)},B=/^\s+|\s+$/g,V=/^[-+]0x[0-9a-f]+$/i,W=/^0b[01]+$/i,X=/^0o[0-7]+$/i,H=parseInt;var Y=function(e){if("number"==typeof e)return e;if($(e))return NaN;if(O(e)){var t="function"==typeof e.valueOf?e.valueOf():e;e=O(t)?t+"":t}if("string"!=typeof e)return 0===e?e:+e;e=e.replace(B,"");var n=W.test(e);return n||X.test(e)?H(e.slice(2),n?2:8):V.test(e)?NaN:+e};var G=function(e){return e?(e=Y(e))===1/0||e===-1/0?17976931348623157e292*(e<0?-1:1):e==e?e:0:0===e?e:0};var K=function(e){var t=G(e),n=t%1;return t==t?n?t-n:t:0},Z=Math.min;var Q=function(e,t){if((e=K(e))<1||e>9007199254740991)return[];var n=4294967295,r=Z(e,4294967295);t=T(t),e-=4294967295;for(var o=_(r,t);++n<e;)t(n);return o};var J=function(){this.__data__=[],this.size=0};var ee=function(e,t){return e===t||e!=e&&t!=t};var te=function(e,t){for(var n=e.length;n--;)if(ee(e[n][0],t))return n;return-1},ne=Array.prototype.splice;var re=function(e){var t=this.__data__,n=te(t,e);return!(n<0)&&(n==t.length-1?t.pop():ne.call(t,n,1),--this.size,!0)};var oe=function(e){var t=this.__data__,n=te(t,e);return n<0?void 0:t[n][1]};var ae=function(e){return te(this.__data__,e)>-1};var ie=function(e,t){var n=this.__data__,r=te(n,e);return r<0?(++this.size,n.push([e,t])):n[r][1]=t,this};function ue(e){var t=-1,n=null==e?0:e.length;for(this.clear();++t<n;){var r=e[t];this.set(r[0],r[1])}}ue.prototype.clear=J,ue.prototype.delete=re,ue.prototype.get=oe,ue.prototype.has=ae,ue.prototype.set=ie;var se=ue;var ce=function(){this.__data__=new se,this.size=0};var le=function(e){var t=this.__data__,n=t.delete(e);return this.size=t.size,n};var fe=function(e){return this.__data__.get(e)};var pe=function(e){return this.__data__.has(e)};var de,he=function(e){if(!O(e))return!1;var t=F(e);return"[object Function]"==t||"[object GeneratorFunction]"==t||"[object AsyncFunction]"==t||"[object Proxy]"==t},ve=I["__core-js_shared__"],me=(de=/[^.]+$/.exec(ve&&ve.keys&&ve.keys.IE_PROTO||""))?"Symbol(src)_1."+de:"";var ge=function(e){return!!me&&me in e},be=Function.prototype.toString;var ye=function(e){if(null!=e){try{return be.call(e)}catch(e){}try{return e+""}catch(e){}}return""},we=/^\[object .+?Constructor\]$/,Ne=Function.prototype,_e=Object.prototype,Ee=Ne.toString,Te=_e.hasOwnProperty,Oe=RegExp("^"+Ee.call(Te).replace(/[\\^$.*+?()[\]{}|]/g,"\\$&").replace(/hasOwnProperty|(function).*?(?=\\\()| for .+?(?=\\\])/g,"$1.*?")+"$");var xe=function(e){return!(!O(e)||ge(e))&&(he(e)?Oe:we).test(ye(e))};var Se=function(e,t){return null==e?void 0:e[t]};var De=function(e,t){var n=Se(e,t);return xe(n)?n:void 0},Ae=De(I,"Map"),Ie=De(Object,"create");var Re=function(){this.__data__=Ie?Ie(null):{},this.size=0};var Ce=function(e){var t=this.has(e)&&delete this.__data__[e];return this.size-=t?1:0,t},ke=Object.prototype.hasOwnProperty;var je=function(e){var t=this.__data__;if(Ie){var n=t[e];return"__lodash_hash_undefined__"===n?void 0:n}return ke.call(t,e)?t[e]:void 0},Ue=Object.prototype.hasOwnProperty;var Pe=function(e){var t=this.__data__;return Ie?void 0!==t[e]:Ue.call(t,e)};var Le=function(e,t){var n=this.__data__;return this.size+=this.has(e)?0:1,n[e]=Ie&&void 0===t?"__lodash_hash_undefined__":t,this};function Me(e){var t=-1,n=null==e?0:e.length;for(this.clear();++t<n;){var r=e[t];this.set(r[0],r[1])}}Me.prototype.clear=Re,Me.prototype.delete=Ce,Me.prototype.get=je,Me.prototype.has=Pe,Me.prototype.set=Le;var qe=Me;var Fe=function(){this.size=0,this.__data__={hash:new qe,map:new(Ae||se),string:new qe}};var ze=function(e){var t=typeof e;return"string"==t||"number"==t||"symbol"==t||"boolean"==t?"__proto__"!==e:null===e};var $e=function(e,t){var n=e.__data__;return ze(t)?n["string"==typeof t?"string":"hash"]:n.map};var Be=function(e){var t=$e(this,e).delete(e);return this.size-=t?1:0,t};var Ve=function(e){return $e(this,e).get(e)};var We=function(e){return $e(this,e).has(e)};var Xe=function(e,t){var n=$e(this,e),r=n.size;return n.set(e,t),this.size+=n.size==r?0:1,this};function He(e){var t=-1,n=null==e?0:e.length;for(this.clear();++t<n;){var r=e[t];this.set(r[0],r[1])}}He.prototype.clear=Fe,He.prototype.delete=Be,He.prototype.get=Ve,He.prototype.has=We,He.prototype.set=Xe;var Ye=He;var Ge=function(e,t){var n=this.__data__;if(n instanceof se){var r=n.__data__;if(!Ae||r.length<199)return r.push([e,t]),this.size=++n.size,this;n=this.__data__=new Ye(r)}return n.set(e,t),this.size=n.size,this};function Ke(e){var t=this.__data__=new se(e);this.size=t.size}Ke.prototype.clear=ce,Ke.prototype.delete=le,Ke.prototype.get=fe,Ke.prototype.has=pe,Ke.prototype.set=Ge;var Ze=Ke,Qe=function(){try{var e=De(Object,"defineProperty");return e({},"",{}),e}catch(e){}}();var Je=function(e,t,n){"__proto__"==t&&Qe?Qe(e,t,{configurable:!0,enumerable:!0,value:n,writable:!0}):e[t]=n};var et=function(e,t,n){(void 0!==n&&!ee(e[t],n)||void 0===n&&!(t in e))&&Je(e,t,n)};var tt=function(e){return function(t,n,r){for(var o=-1,a=Object(t),i=r(t),u=i.length;u--;){var s=i[e?u:++o];if(!1===n(a[s],s,a))break}return t}}(),nt=S((function(e,t){var n=t&&!t.nodeType&&t,r=n&&e&&!e.nodeType&&e,o=r&&r.exports===n?I.Buffer:void 0,a=o?o.allocUnsafe:void 0;e.exports=function(e,t){if(t)return e.slice();var n=e.length,r=a?a(n):new e.constructor(n);return e.copy(r),r}})),rt=I.Uint8Array;var ot=function(e){var t=new e.constructor(e.byteLength);return new rt(t).set(new rt(e)),t};var at=function(e,t){var n=t?ot(e.buffer):e.buffer;return new e.constructor(n,e.byteOffset,e.length)};var it=function(e,t){var n=-1,r=e.length;for(t||(t=Array(r));++n<r;)t[n]=e[n];return t},ut=Object.create,st=function(){function e(){}return function(t){if(!O(t))return{};if(ut)return ut(t);e.prototype=t;var n=new e;return e.prototype=void 0,n}}();var ct=function(e,t){return function(n){return e(t(n))}}(Object.getPrototypeOf,Object),lt=Object.prototype;var ft=function(e){var t=e&&e.constructor;return e===("function"==typeof t&&t.prototype||lt)};var pt=function(e){return"function"!=typeof e.constructor||ft(e)?{}:st(ct(e))};var dt=function(e){return z(e)&&"[object Arguments]"==F(e)},ht=Object.prototype,vt=ht.hasOwnProperty,mt=ht.propertyIsEnumerable,gt=dt(function(){return arguments}())?dt:function(e){return z(e)&&vt.call(e,"callee")&&!mt.call(e,"callee")},bt=Array.isArray;var yt=function(e){return"number"==typeof e&&e>-1&&e%1==0&&e<=9007199254740991};var wt=function(e){return null!=e&&yt(e.length)&&!he(e)};var Nt=function(e){return z(e)&&wt(e)};var _t=function(){return!1},Et=S((function(e,t){var n=t&&!t.nodeType&&t,r=n&&e&&!e.nodeType&&e,o=r&&r.exports===n?I.Buffer:void 0,a=(o?o.isBuffer:void 0)||_t;e.exports=a})),Tt=Function.prototype,Ot=Object.prototype,xt=Tt.toString,St=Ot.hasOwnProperty,Dt=xt.call(Object);var At=function(e){if(!z(e)||"[object Object]"!=F(e))return!1;var t=ct(e);if(null===t)return!0;var n=St.call(t,"constructor")&&t.constructor;return"function"==typeof n&&n instanceof n&&xt.call(n)==Dt},It={};It["[object Float32Array]"]=It["[object Float64Array]"]=It["[object Int8Array]"]=It["[object Int16Array]"]=It["[object Int32Array]"]=It["[object Uint8Array]"]=It["[object Uint8ClampedArray]"]=It["[object Uint16Array]"]=It["[object Uint32Array]"]=!0,It["[object Arguments]"]=It["[object Array]"]=It["[object ArrayBuffer]"]=It["[object Boolean]"]=It["[object DataView]"]=It["[object Date]"]=It["[object Error]"]=It["[object Function]"]=It["[object Map]"]=It["[object Number]"]=It["[object Object]"]=It["[object RegExp]"]=It["[object Set]"]=It["[object String]"]=It["[object WeakMap]"]=!1;var Rt=function(e){return z(e)&&yt(e.length)&&!!It[F(e)]};var Ct=function(e){return function(t){return e(t)}},kt=S((function(e,t){var n=t&&!t.nodeType&&t,r=n&&e&&!e.nodeType&&e,o=r&&r.exports===n&&D.process,a=function(){try{var e=r&&r.require&&r.require("util").types;return e||o&&o.binding&&o.binding("util")}catch(e){}}();e.exports=a})),jt=kt&&kt.isTypedArray,Ut=jt?Ct(jt):Rt;var Pt=function(e,t){if(("constructor"!==t||"function"!=typeof e[t])&&"__proto__"!=t)return e[t]},Lt=Object.prototype.hasOwnProperty;var Mt=function(e,t,n){var r=e[t];Lt.call(e,t)&&ee(r,n)&&(void 0!==n||t in e)||Je(e,t,n)};var qt=function(e,t,n,r){var o=!n;n||(n={});for(var a=-1,i=t.length;++a<i;){var u=t[a],s=r?r(n[u],e[u],u,n,e):void 0;void 0===s&&(s=e[u]),o?Je(n,u,s):Mt(n,u,s)}return n},Ft=/^(?:0|[1-9]\d*)$/;var zt=function(e,t){var n=typeof e;return!!(t=null==t?9007199254740991:t)&&("number"==n||"symbol"!=n&&Ft.test(e))&&e>-1&&e%1==0&&e<t},$t=Object.prototype.hasOwnProperty;var Bt=function(e,t){var n=bt(e),r=!n&&gt(e),o=!n&&!r&&Et(e),a=!n&&!r&&!o&&Ut(e),i=n||r||o||a,u=i?_(e.length,String):[],s=u.length;for(var c in e)!t&&!$t.call(e,c)||i&&("length"==c||o&&("offset"==c||"parent"==c)||a&&("buffer"==c||"byteLength"==c||"byteOffset"==c)||zt(c,s))||u.push(c);return u};var Vt=function(e){var t=[];if(null!=e)for(var n in Object(e))t.push(n);return t},Wt=Object.prototype.hasOwnProperty;var Xt=function(e){if(!O(e))return Vt(e);var t=ft(e),n=[];for(var r in e)("constructor"!=r||!t&&Wt.call(e,r))&&n.push(r);return n};var Ht=function(e){return wt(e)?Bt(e,!0):Xt(e)};var Yt=function(e){return qt(e,Ht(e))};var Gt=function(e,t,n,r,o,a,i){var u=Pt(e,n),s=Pt(t,n),c=i.get(s);if(c)et(e,n,c);else{var l=a?a(u,s,n+"",e,t,i):void 0,f=void 0===l;if(f){var p=bt(s),d=!p&&Et(s),h=!p&&!d&&Ut(s);l=s,p||d||h?bt(u)?l=u:Nt(u)?l=it(u):d?(f=!1,l=nt(s,!0)):h?(f=!1,l=at(s,!0)):l=[]:At(s)||gt(s)?(l=u,gt(u)?l=Yt(u):O(u)&&!he(u)||(l=pt(s))):f=!1}f&&(i.set(s,l),o(l,s,r,a,i),i.delete(s)),et(e,n,l)}};var Kt=function e(t,n,r,o,a){t!==n&&tt(n,(function(i,u){if(a||(a=new Ze),O(i))Gt(t,n,u,r,e,o,a);else{var s=o?o(Pt(t,u),i,u+"",t,n,a):void 0;void 0===s&&(s=i),et(t,u,s)}}),Ht)};var Zt=function(e,t,n){switch(n.length){case 0:return e.call(t);case 1:return e.call(t,n[0]);case 2:return e.call(t,n[0],n[1]);case 3:return e.call(t,n[0],n[1],n[2])}return e.apply(t,n)},Qt=Math.max;var Jt=function(e,t,n){return t=Qt(void 0===t?e.length-1:t,0),function(){for(var r=arguments,o=-1,a=Qt(r.length-t,0),i=Array(a);++o<a;)i[o]=r[t+o];o=-1;for(var u=Array(t+1);++o<t;)u[o]=r[o];return u[t]=n(i),Zt(e,this,u)}};var en=function(e){return function(){return e}},tn=Qe?function(e,t){return Qe(e,"toString",{configurable:!0,enumerable:!1,value:en(t),writable:!0})}:E,nn=Date.now;var rn=function(e){var t=0,n=0;return function(){var r=nn(),o=16-(r-n);if(n=r,o>0){if(++t>=800)return arguments[0]}else t=0;return e.apply(void 0,arguments)}}(tn);var on=function(e,t){return rn(Jt(e,t,E),e+"")};var an=function(e,t,n){if(!O(n))return!1;var r=typeof t;return!!("number"==r?wt(n)&&zt(t,n.length):"string"==r&&t in n)&&ee(n[t],e)};var un=function(e){return on((function(t,n){var r=-1,o=n.length,a=o>1?n[o-1]:void 0,i=o>2?n[2]:void 0;for(a=e.length>3&&"function"==typeof a?(o--,a):void 0,i&&an(n[0],n[1],i)&&(a=o<3?void 0:a,o=1),t=Object(t);++r<o;){var u=n[r];u&&e(t,u,r,a)}return t}))}((function(e,t,n){Kt(e,t,n)})),sn=function(e,t){var n=t.credential,r=void 0!==n.tokenType,o=n,a=n,i={};return r&&un(i,{headers:{Authorization:"".concat(o.tokenType," ").concat(o.accessToken)}}),!r&&a.login&&a.password&&un(i,{headers:{Authorization:"Basic ".concat(s.default.b64encode("".concat(a.login,":").concat(a.password)))}}),un(i,e)},cn=function(){function e(t){c(this,e),p(this,"baseURL",void 0),this.baseURL=t}return f(e,[{key:"create",value:function(e,t,r){var o=u.default.request("MKCOL","".concat(this.baseURL,"/remote.php/dav/files/").concat(e,"/").concat(t),void 0,sn({},{credential:r}));return n.check(o,{"dav create":function(e){return 201===e.status}}),{response:o}}},{key:"delete",value:function(e,t,r){var o=u.default.request("DELETE","".concat(this.baseURL,"/remote.php/dav/files/").concat(e,"/").concat(t),void 0,sn({},{credential:r}));return n.check(o,{"dav delete":function(e){return 204===e.status}}),{response:o}}},{key:"move",value:function(e,t,r,o){var a=u.default.request("MOVE","".concat(this.baseURL,"/remote.php/dav/files/").concat(e,"/").concat(t),void 0,sn({headers:{destination:"/remote.php/dav/files/".concat(e,"/").concat(r)}},{credential:o}));return n.check(a,{"dav move":function(e){return 201===e.status}}),{response:a}}},{key:"upload",value:function(e,t,r,o){var a=u.default.request("PUT","".concat(this.baseURL,"/remote.php/dav/files/").concat(e,"/").concat(t),r,sn({},{credential:o}));return n.check(a,{"dav upload":function(e){return 201===e.status}}),{response:a}}}]),e}(),ln=function(e){v(r,cn);var t=y(r);function r(){return c(this,r),t.apply(this,arguments)}return f(r,[{key:"create",value:function(e,t,r){var o=u.default.request("MKCOL","".concat(this.baseURL,"/remote.php/dav/spaces/").concat(e,"/").concat(t),void 0,sn({},{credential:r}));return n.check(o,{"dav create":function(e){return 201===e.status}}),{response:o}}},{key:"delete",value:function(e,t,r){var o=u.default.request("DELETE","".concat(this.baseURL,"/remote.php/dav/spaces/").concat(e,"/").concat(t),void 0,sn({},{credential:r}));return n.check(o,{"dav delete":function(e){return 204===e.status}}),{response:o}}},{key:"upload",value:function(e,t,r,o){var a=u.default.request("PUT","".concat(this.baseURL,"/remote.php/dav/spaces/").concat(e,"/").concat(t),r,sn({},{credential:o}));return n.check(a,{"dav upload":function(e){return 201===e.status}}),{response:a}}},{key:"move",value:function(e,t,r,o){var a=u.default.request("MOVE","".concat(this.baseURL,"/remote.php/dav/spaces/").concat(e,"/").concat(t),void 0,sn({headers:{destination:"/remote.php/dav/spaces/".concat(e,"/").concat(r)}},{credential:o}));return n.check(a,{"dav move":function(e){return 201===e.status}}),{response:a}}}]),r}(),fn=function(){function e(t){c(this,e),p(this,"baseURL",void 0),this.baseURL=t}return f(e,[{key:"driveInfo",value:function(e,t){return e}}]),e}(),pn=function(e){v(r,fn);var t=y(r);function r(){return c(this,r),t.apply(this,arguments)}return f(r,[{key:"driveInfo",value:function(e,t){var r=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},o=r.selector,a=u.default.request("GET","".concat(this.baseURL,"/graph/v1.0/me/drives"),void 0,sn({},{credential:t}));n.check(a,{"user driveInfo":function(e){return 200===e.status}});var i=a.json(o);return i||e}}]),r}(),dn={entityMap:{lt:"<",gt:">",amp:"&",quot:'"',apos:"'",Agrave:"À",Aacute:"Á",Acirc:"Â",Atilde:"Ã",Auml:"Ä",Aring:"Å",AElig:"Æ",Ccedil:"Ç",Egrave:"È",Eacute:"É",Ecirc:"Ê",Euml:"Ë",Igrave:"Ì",Iacute:"Í",Icirc:"Î",Iuml:"Ï",ETH:"Ð",Ntilde:"Ñ",Ograve:"Ò",Oacute:"Ó",Ocirc:"Ô",Otilde:"Õ",Ouml:"Ö",Oslash:"Ø",Ugrave:"Ù",Uacute:"Ú",Ucirc:"Û",Uuml:"Ü",Yacute:"Ý",THORN:"Þ",szlig:"ß",agrave:"à",aacute:"á",acirc:"â",atilde:"ã",auml:"ä",aring:"å",aelig:"æ",ccedil:"ç",egrave:"è",eacute:"é",ecirc:"ê",euml:"ë",igrave:"ì",iacute:"í",icirc:"î",iuml:"ï",eth:"ð",ntilde:"ñ",ograve:"ò",oacute:"ó",ocirc:"ô",otilde:"õ",ouml:"ö",oslash:"ø",ugrave:"ù",uacute:"ú",ucirc:"û",uuml:"ü",yacute:"ý",thorn:"þ",yuml:"ÿ",nbsp:" ",iexcl:"¡",cent:"¢",pound:"£",curren:"¤",yen:"¥",brvbar:"¦",sect:"§",uml:"¨",copy:"©",ordf:"ª",laquo:"«",not:"¬",shy:"­­",reg:"®",macr:"¯",deg:"°",plusmn:"±",sup2:"²",sup3:"³",acute:"´",micro:"µ",para:"¶",middot:"·",cedil:"¸",sup1:"¹",ordm:"º",raquo:"»",frac14:"¼",frac12:"½",frac34:"¾",iquest:"¿",times:"×",divide:"÷",forall:"∀",part:"∂",exist:"∃",empty:"∅",nabla:"∇",isin:"∈",notin:"∉",ni:"∋",prod:"∏",sum:"∑",minus:"−",lowast:"∗",radic:"√",prop:"∝",infin:"∞",ang:"∠",and:"∧",or:"∨",cap:"∩",cup:"∪",int:"∫",there4:"∴",sim:"∼",cong:"≅",asymp:"≈",ne:"≠",equiv:"≡",le:"≤",ge:"≥",sub:"⊂",sup:"⊃",nsub:"⊄",sube:"⊆",supe:"⊇",oplus:"⊕",otimes:"⊗",perp:"⊥",sdot:"⋅",Alpha:"Α",Beta:"Β",Gamma:"Γ",Delta:"Δ",Epsilon:"Ε",Zeta:"Ζ",Eta:"Η",Theta:"Θ",Iota:"Ι",Kappa:"Κ",Lambda:"Λ",Mu:"Μ",Nu:"Ν",Xi:"Ξ",Omicron:"Ο",Pi:"Π",Rho:"Ρ",Sigma:"Σ",Tau:"Τ",Upsilon:"Υ",Phi:"Φ",Chi:"Χ",Psi:"Ψ",Omega:"Ω",alpha:"α",beta:"β",gamma:"γ",delta:"δ",epsilon:"ε",zeta:"ζ",eta:"η",theta:"θ",iota:"ι",kappa:"κ",lambda:"λ",mu:"μ",nu:"ν",xi:"ξ",omicron:"ο",pi:"π",rho:"ρ",sigmaf:"ς",sigma:"σ",tau:"τ",upsilon:"υ",phi:"φ",chi:"χ",psi:"ψ",omega:"ω",thetasym:"ϑ",upsih:"ϒ",piv:"ϖ",OElig:"Œ",oelig:"œ",Scaron:"Š",scaron:"š",Yuml:"Ÿ",fnof:"ƒ",circ:"ˆ",tilde:"˜",ensp:" ",emsp:" ",thinsp:" ",zwnj:"‌",zwj:"‍",lrm:"‎",rlm:"‏",ndash:"–",mdash:"—",lsquo:"‘",rsquo:"’",sbquo:"‚",ldquo:"“",rdquo:"”",bdquo:"„",dagger:"†",Dagger:"‡",bull:"•",hellip:"…",permil:"‰",prime:"′",Prime:"″",lsaquo:"‹",rsaquo:"›",oline:"‾",euro:"€",trade:"™",larr:"←",uarr:"↑",rarr:"→",darr:"↓",harr:"↔",crarr:"↵",lceil:"⌈",rceil:"⌉",lfloor:"⌊",rfloor:"⌋",loz:"◊",spades:"♠",clubs:"♣",hearts:"♥",diams:"♦"}},hn=/[A-Z_a-z\xC0-\xD6\xD8-\xF6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD]/,vn=new RegExp("[\\-\\.0-9"+hn.source.slice(1,-1)+"\\u00B7\\u0300-\\u036F\\u203F-\\u2040]"),mn=new RegExp("^"+hn.source+vn.source+"*(?::"+hn.source+vn.source+"*)?$");function gn(){}function bn(e,t){return t.lineNumber=e.lineNumber,t.columnNumber=e.columnNumber,t}function yn(e,t,n,r,o,a){for(var i,u=++t,s=0;;){var c=e.charAt(u);switch(c){case"=":if(1===s)i=e.slice(t,u),s=3;else{if(2!==s)throw new Error("attribute equal must after attrName");s=3}break;case"'":case'"':if(3===s||1===s){if(1===s&&(a.warning('attribute value must after "="'),i=e.slice(t,u)),t=u+1,!((u=e.indexOf(c,t))>0))throw new Error("attribute value no end '"+c+"' match");l=e.slice(t,u).replace(/&#?\w+;/g,o),n.add(i,l,t-1),s=5}else{if(4!=s)throw new Error('attribute value must after "="');l=e.slice(t,u).replace(/&#?\w+;/g,o),n.add(i,l,t),a.warning('attribute "'+i+'" missed start quot('+c+")!!"),t=u+1,s=5}break;case"/":switch(s){case 0:n.setTagName(e.slice(t,u));case 5:case 6:case 7:s=7,n.closed=!0;case 4:case 1:case 2:break;default:throw new Error("attribute invalid close char('/')")}break;case"":return a.error("unexpected end of input"),0==s&&n.setTagName(e.slice(t,u)),u;case">":switch(s){case 0:n.setTagName(e.slice(t,u));case 5:case 6:case 7:break;case 4:case 1:"/"===(l=e.slice(t,u)).slice(-1)&&(n.closed=!0,l=l.slice(0,-1));case 2:2===s&&(l=i),4==s?(a.warning('attribute "'+l+'" missed quot(")!!'),n.add(i,l.replace(/&#?\w+;/g,o),t)):("http://www.w3.org/1999/xhtml"===r[""]&&l.match(/^(?:disabled|checked|selected)$/i)||a.warning('attribute "'+l+'" missed value!! "'+l+'" instead!!'),n.add(l,l,t));break;case 3:throw new Error("attribute value missed!!")}return u;case"":c=" ";default:if(c<=" ")switch(s){case 0:n.setTagName(e.slice(t,u)),s=6;break;case 1:i=e.slice(t,u),s=2;break;case 4:var l=e.slice(t,u).replace(/&#?\w+;/g,o);a.warning('attribute "'+l+'" missed quot(")!!'),n.add(i,l,t);case 5:s=6}else switch(s){case 2:n.tagName,"http://www.w3.org/1999/xhtml"===r[""]&&i.match(/^(?:disabled|checked|selected)$/i)||a.warning('attribute "'+i+'" missed value!! "'+i+'" instead2!!'),n.add(i,i,t),t=u,s=1;break;case 5:a.warning('attribute space is required"'+i+'"!!');case 6:s=1,t=u;break;case 3:s=4,t=u;break;case 7:throw new Error("elements closed character '/' and '>' must be connected to")}}u++}}function wn(e,t,n){for(var r=e.tagName,o=null,a=e.length;a--;){var i=e[a],u=i.qName,s=i.value;if((p=u.indexOf(":"))>0)var c=i.prefix=u.slice(0,p),l=u.slice(p+1),f="xmlns"===c&&l;else l=u,c=null,f="xmlns"===u&&"";i.localName=l,!1!==f&&(null==o&&(o={},En(n,n={})),n[f]=o[f]=s,i.uri="http://www.w3.org/2000/xmlns/",t.startPrefixMapping(f,s))}for(a=e.length;a--;){(c=(i=e[a]).prefix)&&("xml"===c&&(i.uri="http://www.w3.org/XML/1998/namespace"),"xmlns"!==c&&(i.uri=n[c||""]))}var p;(p=r.indexOf(":"))>0?(c=e.prefix=r.slice(0,p),l=e.localName=r.slice(p+1)):(c=null,l=e.localName=r);var d=e.uri=n[c||""];if(t.startElement(d,l,r,e),!e.closed)return e.currentNSMap=n,e.localNSMap=o,!0;if(t.endElement(d,l,r),o)for(c in o)t.endPrefixMapping(c)}function Nn(e,t,n,r,o){if(/^(?:script|textarea)$/i.test(n)){var a=e.indexOf("</"+n+">",t),i=e.substring(t+1,a);if(/[&<]/.test(i))return/^script$/i.test(n)?(o.characters(i,0,i.length),a):(i=i.replace(/&#?\w+;/g,r),o.characters(i,0,i.length),a)}return t+1}function _n(e,t,n,r){var o=r[n];return null==o&&((o=e.lastIndexOf("</"+n+">"))<t&&(o=e.lastIndexOf("</"+n)),r[n]=o),o<t}function En(e,t){for(var n in e)t[n]=e[n]}function Tn(e,t,n,r){switch(e.charAt(t+2)){case"-":return"-"===e.charAt(t+3)?(o=e.indexOf("--\x3e",t+4))>t?(n.comment(e,t+4,o-t-4),o+3):(r.error("Unclosed comment"),-1):-1;default:if("CDATA["==e.substr(t+3,6)){var o=e.indexOf("]]>",t+9);return n.startCDATA(),n.characters(e,t+9,o-t-9),n.endCDATA(),o+3}var a=function(e,t){var n,r=[],o=/'[^']+'|"[^"]+"|[^\s<>\/=]+=?|(\/?\s*>|<)/g;o.lastIndex=t,o.exec(e);for(;n=o.exec(e);)if(r.push(n),n[1])return r}(e,t),i=a.length;if(i>1&&/!doctype/i.test(a[0][0])){var u=a[1][0],s=!1,c=!1;i>3&&(/^public$/i.test(a[2][0])?(s=a[3][0],c=i>4&&a[4][0]):/^system$/i.test(a[2][0])&&(c=a[3][0]));var l=a[i-1];return n.startDTD(u,s&&s.replace(/^(['"])(.*?)\1$/,"$2"),c&&c.replace(/^(['"])(.*?)\1$/,"$2")),n.endDTD(),l.index+l[0].length}}return-1}function On(e,t,n){var r=e.indexOf("?>",t);if(r){var o=e.substring(t,r).match(/^<\?(\S*)\s*([\s\S]*?)\s*$/);return o?(o[0].length,n.processingInstruction(o[1],o[2]),r+2):-1}return-1}function xn(e){}gn.prototype={parse:function(e,t,n){var r=this.domBuilder;r.startDocument(),En(t,t={}),function(e,t,n,r,o){function a(e){var t=e.slice(1,-1);return t in n?n[t]:"#"===t.charAt(0)?function(e){if(e>65535){var t=55296+((e-=65536)>>10),n=56320+(1023&e);return String.fromCharCode(t,n)}return String.fromCharCode(e)}(parseInt(t.substr(1).replace("x","0x"))):(o.error("entity not found:"+e),e)}function i(t){if(t>h){var n=e.substring(h,t).replace(/&#?\w+;/g,a);f&&u(h),r.characters(n,0,t-h),h=t}}function u(t,n){for(;t>=c&&(n=l.exec(e));)s=n.index,c=s+n[0].length,f.lineNumber++;f.columnNumber=t-s+1}var s=0,c=0,l=/.*(?:\r\n?|\n)|.*$/g,f=r.locator,p=[{currentNSMap:t}],d={},h=0;for(;;){try{var v=e.indexOf("<",h);if(v<0){if(!e.substr(h).match(/^\s*$/)){var m=r.doc,g=m.createTextNode(e.substr(h));m.appendChild(g),r.currentElement=g}return}switch(v>h&&i(v),e.charAt(v+1)){case"/":var b=e.indexOf(">",v+3),y=e.substring(v+2,b),w=p.pop();b<0?(y=e.substring(v+2).replace(/[\s<].*/,""),o.error("end tag name: "+y+" is not complete:"+w.tagName),b=v+1+y.length):y.match(/\s</)&&(y=y.replace(/[\s<].*/,""),o.error("end tag name: "+y+" maybe not complete"),b=v+1+y.length);var N=w.localNSMap,_=w.tagName==y;if(_||w.tagName&&w.tagName.toLowerCase()==y.toLowerCase()){if(r.endElement(w.uri,w.localName,y),N)for(var E in N)r.endPrefixMapping(E);_||o.fatalError("end tag name: "+y+" is not match the current start tagName:"+w.tagName)}else p.push(w);b++;break;case"?":f&&u(v),b=On(e,v,r);break;case"!":f&&u(v),b=Tn(e,v,r,o);break;default:f&&u(v);var T=new xn,O=p[p.length-1].currentNSMap,x=(b=yn(e,v,T,O,a,o),T.length);if(!T.closed&&_n(e,b,T.tagName,d)&&(T.closed=!0,n.nbsp||o.warning("unclosed xml attribute")),f&&x){for(var S=bn(f,{}),D=0;D<x;D++){var A=T[D];u(A.offset),A.locator=bn(f,{})}r.locator=S,wn(T,r,O)&&p.push(T),r.locator=f}else wn(T,r,O)&&p.push(T);"http://www.w3.org/1999/xhtml"!==T.uri||T.closed?b++:b=Nn(e,b,T.tagName,a,r)}}catch(e){o.error("element parse error: "+e),b=-1}b>h?h=b:i(Math.max(v,h)+1)}}(e,t,n,r,this.errorHandler),r.endDocument()}},xn.prototype={setTagName:function(e){if(!mn.test(e))throw new Error("invalid tagName:"+e);this.tagName=e},add:function(e,t,n){if(!mn.test(e))throw new Error("invalid attribute:"+e);this[this.length++]={qName:e,value:t,offset:n}},length:0,getLocalName:function(e){return this[e].localName},getLocator:function(e){return this[e].locator},getQName:function(e){return this[e].qName},getURI:function(e){return this[e].uri},getValue:function(e){return this[e].value}};var Sn={XMLReader:gn};function Dn(e,t){for(var n in e)t[n]=e[n]}function An(e,t){var n=e.prototype;if(!(n instanceof t)){function r(){}r.prototype=t.prototype,Dn(n,r=new r),e.prototype=n=r}n.constructor!=e&&("function"!=typeof e&&console.error("unknow Class:"+e),n.constructor=e)}var In={},Rn=In.ELEMENT_NODE=1,Cn=In.ATTRIBUTE_NODE=2,kn=In.TEXT_NODE=3,jn=In.CDATA_SECTION_NODE=4,Un=In.ENTITY_REFERENCE_NODE=5,Pn=In.ENTITY_NODE=6,Ln=In.PROCESSING_INSTRUCTION_NODE=7,Mn=In.COMMENT_NODE=8,qn=In.DOCUMENT_NODE=9,Fn=In.DOCUMENT_TYPE_NODE=10,zn=In.DOCUMENT_FRAGMENT_NODE=11,$n=In.NOTATION_NODE=12,Bn={},Vn={};Bn.INDEX_SIZE_ERR=(Vn[1]="Index size error",1),Bn.DOMSTRING_SIZE_ERR=(Vn[2]="DOMString size error",2);var Wn=Bn.HIERARCHY_REQUEST_ERR=(Vn[3]="Hierarchy request error",3);Bn.WRONG_DOCUMENT_ERR=(Vn[4]="Wrong document",4),Bn.INVALID_CHARACTER_ERR=(Vn[5]="Invalid character",5),Bn.NO_DATA_ALLOWED_ERR=(Vn[6]="No data allowed",6),Bn.NO_MODIFICATION_ALLOWED_ERR=(Vn[7]="No modification allowed",7);var Xn=Bn.NOT_FOUND_ERR=(Vn[8]="Not found",8);Bn.NOT_SUPPORTED_ERR=(Vn[9]="Not supported",9);var Hn=Bn.INUSE_ATTRIBUTE_ERR=(Vn[10]="Attribute in use",10);function Yn(e,t){if(t instanceof Error)var n=t;else n=this,Error.call(this,Vn[e]),this.message=Vn[e],Error.captureStackTrace&&Error.captureStackTrace(this,Yn);return n.code=e,t&&(this.message=this.message+": "+t),n}function Gn(){}function Kn(e,t){this._node=e,this._refresh=t,Zn(this)}function Zn(e){var t=e._node._inc||e._node.ownerDocument._inc;if(e._inc!=t){var n=e._refresh(e._node);Sr(e,"length",n.length),Dn(n,e),e._inc=t}}function Qn(){}function Jn(e,t){for(var n=e.length;n--;)if(e[n]===t)return n}function er(e,t,n,r){if(r?t[Jn(t,r)]=n:t[t.length++]=n,e){n.ownerElement=e;var o=e.ownerDocument;o&&(r&&ur(o,e,r),function(e,t,n){e&&e._inc++,"http://www.w3.org/2000/xmlns/"==n.namespaceURI&&(t._nsMap[n.prefix?n.localName:""]=n.value)}(o,e,n))}}function tr(e,t,n){var r=Jn(t,n);if(!(r>=0))throw Yn(Xn,new Error(e.tagName+"@"+n));for(var o=t.length-1;r<o;)t[r]=t[++r];if(t.length=o,e){var a=e.ownerDocument;a&&(ur(a,e,n),n.ownerElement=null)}}function nr(e){if(this._features={},e)for(var t in e)this._features=e[t]}function rr(){}function or(e){return("<"==e?"&lt;":">"==e&&"&gt;")||"&"==e&&"&amp;"||'"'==e&&"&quot;"||"&#"+e.charCodeAt()+";"}function ar(e,t){if(t(e))return!0;if(e=e.firstChild)do{if(ar(e,t))return!0}while(e=e.nextSibling)}function ir(){}function ur(e,t,n,r){e&&e._inc++,"http://www.w3.org/2000/xmlns/"==n.namespaceURI&&delete t._nsMap[n.prefix?n.localName:""]}function sr(e,t,n){if(e&&e._inc){e._inc++;var r=t.childNodes;if(n)r[r.length++]=n;else{for(var o=t.firstChild,a=0;o;)r[a++]=o,o=o.nextSibling;r.length=a}}}function cr(e,t){var n=t.previousSibling,r=t.nextSibling;return n?n.nextSibling=r:e.firstChild=r,r?r.previousSibling=n:e.lastChild=n,sr(e.ownerDocument,e),t}function lr(e,t,n){var r=t.parentNode;if(r&&r.removeChild(t),t.nodeType===zn){var o=t.firstChild;if(null==o)return t;var a=t.lastChild}else o=a=t;var i=n?n.previousSibling:e.lastChild;o.previousSibling=i,a.nextSibling=n,i?i.nextSibling=o:e.firstChild=o,null==n?e.lastChild=a:n.previousSibling=a;do{o.parentNode=e}while(o!==a&&(o=o.nextSibling));return sr(e.ownerDocument||e,e),t.nodeType==zn&&(t.firstChild=t.lastChild=null),t}function fr(){this._nsMap={}}function pr(){}function dr(){}function hr(){}function vr(){}function mr(){}function gr(){}function br(){}function yr(){}function wr(){}function Nr(){}function _r(){}function Er(){}function Tr(e,t){var n=[],r=9==this.nodeType&&this.documentElement||this,o=r.prefix,a=r.namespaceURI;if(a&&null==o&&null==(o=r.lookupPrefix(a)))var i=[{namespace:a,prefix:null}];return xr(this,n,e,t,i),n.join("")}function Or(e,t,n){var r=e.prefix||"",o=e.namespaceURI;if(!r&&!o)return!1;if("xml"===r&&"http://www.w3.org/XML/1998/namespace"===o||"http://www.w3.org/2000/xmlns/"==o)return!1;for(var a=n.length;a--;){var i=n[a];if(i.prefix==r)return i.namespace!=o}return!0}function xr(e,t,n,r,o){if(r){if(!(e=r(e)))return;if("string"==typeof e)return void t.push(e)}switch(e.nodeType){case Rn:o||(o=[]),o.length;var a=e.attributes,i=a.length,u=e.firstChild,s=e.tagName;n="http://www.w3.org/1999/xhtml"===e.namespaceURI||n,t.push("<",s);for(var c=0;c<i;c++){"xmlns"==(l=a.item(c)).prefix?o.push({prefix:l.localName,namespace:l.value}):"xmlns"==l.nodeName&&o.push({prefix:"",namespace:l.value})}for(c=0;c<i;c++){var l;if(Or(l=a.item(c),0,o)){var f=l.prefix||"",p=l.namespaceURI,d=f?" xmlns:"+f:" xmlns";t.push(d,'="',p,'"'),o.push({prefix:f,namespace:p})}xr(l,t,n,r,o)}if(Or(e,0,o)){f=e.prefix||"",p=e.namespaceURI,d=f?" xmlns:"+f:" xmlns";t.push(d,'="',p,'"'),o.push({prefix:f,namespace:p})}if(u||n&&!/^(?:meta|link|img|br|hr|input)$/i.test(s)){if(t.push(">"),n&&/^script$/i.test(s))for(;u;)u.data?t.push(u.data):xr(u,t,n,r,o),u=u.nextSibling;else for(;u;)xr(u,t,n,r,o),u=u.nextSibling;t.push("</",s,">")}else t.push("/>");return;case qn:case zn:for(u=e.firstChild;u;)xr(u,t,n,r,o),u=u.nextSibling;return;case Cn:return t.push(" ",e.name,'="',e.value.replace(/[<&"]/g,or),'"');case kn:return t.push(e.data.replace(/[<&]/g,or));case jn:return t.push("<![CDATA[",e.data,"]]>");case Mn:return t.push("\x3c!--",e.data,"--\x3e");case Fn:var h=e.publicId,v=e.systemId;if(t.push("<!DOCTYPE ",e.name),h)t.push(' PUBLIC "',h),v&&"."!=v&&t.push('" "',v),t.push('">');else if(v&&"."!=v)t.push(' SYSTEM "',v,'">');else{var m=e.internalSubset;m&&t.push(" [",m,"]"),t.push(">")}return;case Ln:return t.push("<?",e.target," ",e.data,"?>");case Un:return t.push("&",e.nodeName,";");default:t.push("??",e.nodeName)}}function Sr(e,t,n){e[t]=n}Bn.INVALID_STATE_ERR=(Vn[11]="Invalid state",11),Bn.SYNTAX_ERR=(Vn[12]="Syntax error",12),Bn.INVALID_MODIFICATION_ERR=(Vn[13]="Invalid modification",13),Bn.NAMESPACE_ERR=(Vn[14]="Invalid namespace",14),Bn.INVALID_ACCESS_ERR=(Vn[15]="Invalid access",15),Yn.prototype=Error.prototype,Dn(Bn,Yn),Gn.prototype={length:0,item:function(e){return this[e]||null},toString:function(e,t){for(var n=[],r=0;r<this.length;r++)xr(this[r],n,e,t);return n.join("")}},Kn.prototype.item=function(e){return Zn(this),this[e]},An(Kn,Gn),Qn.prototype={length:0,item:Gn.prototype.item,getNamedItem:function(e){for(var t=this.length;t--;){var n=this[t];if(n.nodeName==e)return n}},setNamedItem:function(e){var t=e.ownerElement;if(t&&t!=this._ownerElement)throw new Yn(Hn);var n=this.getNamedItem(e.nodeName);return er(this._ownerElement,this,e,n),n},setNamedItemNS:function(e){var t,n=e.ownerElement;if(n&&n!=this._ownerElement)throw new Yn(Hn);return t=this.getNamedItemNS(e.namespaceURI,e.localName),er(this._ownerElement,this,e,t),t},removeNamedItem:function(e){var t=this.getNamedItem(e);return tr(this._ownerElement,this,t),t},removeNamedItemNS:function(e,t){var n=this.getNamedItemNS(e,t);return tr(this._ownerElement,this,n),n},getNamedItemNS:function(e,t){for(var n=this.length;n--;){var r=this[n];if(r.localName==t&&r.namespaceURI==e)return r}return null}},nr.prototype={hasFeature:function(e,t){var n=this._features[e.toLowerCase()];return!(!n||t&&!(t in n))},createDocument:function(e,t,n){var r=new ir;if(r.implementation=this,r.childNodes=new Gn,r.doctype=n,n&&r.appendChild(n),t){var o=r.createElementNS(e,t);r.appendChild(o)}return r},createDocumentType:function(e,t,n){var r=new gr;return r.name=e,r.nodeName=e,r.publicId=t,r.systemId=n,r}},rr.prototype={firstChild:null,lastChild:null,previousSibling:null,nextSibling:null,attributes:null,parentNode:null,childNodes:null,ownerDocument:null,nodeValue:null,namespaceURI:null,prefix:null,localName:null,insertBefore:function(e,t){return lr(this,e,t)},replaceChild:function(e,t){this.insertBefore(e,t),t&&this.removeChild(t)},removeChild:function(e){return cr(this,e)},appendChild:function(e){return this.insertBefore(e,null)},hasChildNodes:function(){return null!=this.firstChild},cloneNode:function(e){return function e(t,n,r){var o=new n.constructor;for(var a in n){var i=n[a];"object"!=typeof i&&i!=o[a]&&(o[a]=i)}n.childNodes&&(o.childNodes=new Gn);switch(o.ownerDocument=t,o.nodeType){case Rn:var u=n.attributes,s=o.attributes=new Qn,c=u.length;s._ownerElement=o;for(var l=0;l<c;l++)o.setAttributeNode(e(t,u.item(l),!0));break;case Cn:r=!0}if(r)for(var f=n.firstChild;f;)o.appendChild(e(t,f,r)),f=f.nextSibling;return o}(this.ownerDocument||this,this,e)},normalize:function(){for(var e=this.firstChild;e;){var t=e.nextSibling;t&&t.nodeType==kn&&e.nodeType==kn?(this.removeChild(t),e.appendData(t.data)):(e.normalize(),e=t)}},isSupported:function(e,t){return this.ownerDocument.implementation.hasFeature(e,t)},hasAttributes:function(){return this.attributes.length>0},lookupPrefix:function(e){for(var t=this;t;){var n=t._nsMap;if(n)for(var r in n)if(n[r]==e)return r;t=t.nodeType==Cn?t.ownerDocument:t.parentNode}return null},lookupNamespaceURI:function(e){for(var t=this;t;){var n=t._nsMap;if(n&&e in n)return n[e];t=t.nodeType==Cn?t.ownerDocument:t.parentNode}return null},isDefaultNamespace:function(e){return null==this.lookupPrefix(e)}},Dn(In,rr),Dn(In,rr.prototype),ir.prototype={nodeName:"#document",nodeType:qn,doctype:null,documentElement:null,_inc:1,insertBefore:function(e,t){if(e.nodeType==zn){for(var n=e.firstChild;n;){var r=n.nextSibling;this.insertBefore(n,t),n=r}return e}return null==this.documentElement&&e.nodeType==Rn&&(this.documentElement=e),lr(this,e,t),e.ownerDocument=this,e},removeChild:function(e){return this.documentElement==e&&(this.documentElement=null),cr(this,e)},importNode:function(e,t){return function e(t,n,r){var o;switch(n.nodeType){case Rn:(o=n.cloneNode(!1)).ownerDocument=t;case zn:break;case Cn:r=!0}o||(o=n.cloneNode(!1));if(o.ownerDocument=t,o.parentNode=null,r)for(var a=n.firstChild;a;)o.appendChild(e(t,a,r)),a=a.nextSibling;return o}(this,e,t)},getElementById:function(e){var t=null;return ar(this.documentElement,(function(n){if(n.nodeType==Rn&&n.getAttribute("id")==e)return t=n,!0})),t},getElementsByClassName:function(e){var t=new RegExp("(^|\\s)"+e+"(\\s|$)");return new Kn(this,(function(e){var n=[];return ar(e.documentElement,(function(r){r!==e&&r.nodeType==Rn&&t.test(r.getAttribute("class"))&&n.push(r)})),n}))},createElement:function(e){var t=new fr;return t.ownerDocument=this,t.nodeName=e,t.tagName=e,t.childNodes=new Gn,(t.attributes=new Qn)._ownerElement=t,t},createDocumentFragment:function(){var e=new Nr;return e.ownerDocument=this,e.childNodes=new Gn,e},createTextNode:function(e){var t=new hr;return t.ownerDocument=this,t.appendData(e),t},createComment:function(e){var t=new vr;return t.ownerDocument=this,t.appendData(e),t},createCDATASection:function(e){var t=new mr;return t.ownerDocument=this,t.appendData(e),t},createProcessingInstruction:function(e,t){var n=new _r;return n.ownerDocument=this,n.tagName=n.target=e,n.nodeValue=n.data=t,n},createAttribute:function(e){var t=new pr;return t.ownerDocument=this,t.name=e,t.nodeName=e,t.localName=e,t.specified=!0,t},createEntityReference:function(e){var t=new wr;return t.ownerDocument=this,t.nodeName=e,t},createElementNS:function(e,t){var n=new fr,r=t.split(":"),o=n.attributes=new Qn;return n.childNodes=new Gn,n.ownerDocument=this,n.nodeName=t,n.tagName=t,n.namespaceURI=e,2==r.length?(n.prefix=r[0],n.localName=r[1]):n.localName=t,o._ownerElement=n,n},createAttributeNS:function(e,t){var n=new pr,r=t.split(":");return n.ownerDocument=this,n.nodeName=t,n.name=t,n.namespaceURI=e,n.specified=!0,2==r.length?(n.prefix=r[0],n.localName=r[1]):n.localName=t,n}},An(ir,rr),fr.prototype={nodeType:Rn,hasAttribute:function(e){return null!=this.getAttributeNode(e)},getAttribute:function(e){var t=this.getAttributeNode(e);return t&&t.value||""},getAttributeNode:function(e){return this.attributes.getNamedItem(e)},setAttribute:function(e,t){var n=this.ownerDocument.createAttribute(e);n.value=n.nodeValue=""+t,this.setAttributeNode(n)},removeAttribute:function(e){var t=this.getAttributeNode(e);t&&this.removeAttributeNode(t)},appendChild:function(e){return e.nodeType===zn?this.insertBefore(e,null):function(e,t){var n=t.parentNode;if(n){var r=e.lastChild;n.removeChild(t);r=e.lastChild}return r=e.lastChild,t.parentNode=e,t.previousSibling=r,t.nextSibling=null,r?r.nextSibling=t:e.firstChild=t,e.lastChild=t,sr(e.ownerDocument,e,t),t}(this,e)},setAttributeNode:function(e){return this.attributes.setNamedItem(e)},setAttributeNodeNS:function(e){return this.attributes.setNamedItemNS(e)},removeAttributeNode:function(e){return this.attributes.removeNamedItem(e.nodeName)},removeAttributeNS:function(e,t){var n=this.getAttributeNodeNS(e,t);n&&this.removeAttributeNode(n)},hasAttributeNS:function(e,t){return null!=this.getAttributeNodeNS(e,t)},getAttributeNS:function(e,t){var n=this.getAttributeNodeNS(e,t);return n&&n.value||""},setAttributeNS:function(e,t,n){var r=this.ownerDocument.createAttributeNS(e,t);r.value=r.nodeValue=""+n,this.setAttributeNode(r)},getAttributeNodeNS:function(e,t){return this.attributes.getNamedItemNS(e,t)},getElementsByTagName:function(e){return new Kn(this,(function(t){var n=[];return ar(t,(function(r){r===t||r.nodeType!=Rn||"*"!==e&&r.tagName!=e||n.push(r)})),n}))},getElementsByTagNameNS:function(e,t){return new Kn(this,(function(n){var r=[];return ar(n,(function(o){o===n||o.nodeType!==Rn||"*"!==e&&o.namespaceURI!==e||"*"!==t&&o.localName!=t||r.push(o)})),r}))}},ir.prototype.getElementsByTagName=fr.prototype.getElementsByTagName,ir.prototype.getElementsByTagNameNS=fr.prototype.getElementsByTagNameNS,An(fr,rr),pr.prototype.nodeType=Cn,An(pr,rr),dr.prototype={data:"",substringData:function(e,t){return this.data.substring(e,e+t)},appendData:function(e){e=this.data+e,this.nodeValue=this.data=e,this.length=e.length},insertData:function(e,t){this.replaceData(e,0,t)},appendChild:function(e){throw new Error(Vn[Wn])},deleteData:function(e,t){this.replaceData(e,t,"")},replaceData:function(e,t,n){n=this.data.substring(0,e)+n+this.data.substring(e+t),this.nodeValue=this.data=n,this.length=n.length}},An(dr,rr),hr.prototype={nodeName:"#text",nodeType:kn,splitText:function(e){var t=this.data,n=t.substring(e);t=t.substring(0,e),this.data=this.nodeValue=t,this.length=t.length;var r=this.ownerDocument.createTextNode(n);return this.parentNode&&this.parentNode.insertBefore(r,this.nextSibling),r}},An(hr,dr),vr.prototype={nodeName:"#comment",nodeType:Mn},An(vr,dr),mr.prototype={nodeName:"#cdata-section",nodeType:jn},An(mr,dr),gr.prototype.nodeType=Fn,An(gr,rr),br.prototype.nodeType=$n,An(br,rr),yr.prototype.nodeType=Pn,An(yr,rr),wr.prototype.nodeType=Un,An(wr,rr),Nr.prototype.nodeName="#document-fragment",Nr.prototype.nodeType=zn,An(Nr,rr),_r.prototype.nodeType=Ln,An(_r,rr),Er.prototype.serializeToString=function(e,t,n){return Tr.call(e,t,n)},rr.prototype.toString=Tr;try{if(Object.defineProperty){Object.defineProperty(Kn.prototype,"length",{get:function(){return Zn(this),this.$$length}}),Object.defineProperty(rr.prototype,"textContent",{get:function(){return function e(t){switch(t.nodeType){case Rn:case zn:var n=[];for(t=t.firstChild;t;)7!==t.nodeType&&8!==t.nodeType&&n.push(e(t)),t=t.nextSibling;return n.join("");default:return t.nodeValue}}(this)},set:function(e){switch(this.nodeType){case Rn:case zn:for(;this.firstChild;)this.removeChild(this.firstChild);(e||String(e))&&this.appendChild(this.ownerDocument.createTextNode(e));break;default:this.data=e,this.value=e,this.nodeValue=e}}}),Sr=function(e,t,n){e["$$"+t]=n}}}catch(e){}var Dr={Node:rr,DOMImplementation:nr,XMLSerializer:Er},Ar=S((function(e,t){function n(e){this.options=e||{locator:{}}}function r(){this.cdata=!1}function o(e,t){t.lineNumber=e.lineNumber,t.columnNumber=e.columnNumber}function a(e){if(e)return"\n@"+(e.systemId||"")+"#[line:"+e.lineNumber+",col:"+e.columnNumber+"]"}function i(e,t,n){return"string"==typeof e?e.substr(t,n):e.length>=t+n||t?new java.lang.String(e,t,n)+"":e}function u(e,t){e.currentElement?e.currentElement.appendChild(t):e.doc.appendChild(t)}n.prototype.parseFromString=function(e,t){var n=this.options,o=new s,i=n.domBuilder||new r,u=n.errorHandler,c=n.locator,l=n.xmlns||{},f=/\/x?html?$/.test(t),p=f?dn.entityMap:{lt:"<",gt:">",amp:"&",quot:'"',apos:"'"};return c&&i.setDocumentLocator(c),o.errorHandler=function(e,t,n){if(!e){if(t instanceof r)return t;e=t}var o={},i=e instanceof Function;function u(t){var r=e[t];!r&&i&&(r=2==e.length?function(n){e(t,n)}:e),o[t]=r&&function(e){r("[xmldom "+t+"]\t"+e+a(n))}||function(){}}return n=n||{},u("warning"),u("error"),u("fatalError"),o}(u,i,c),o.domBuilder=n.domBuilder||i,f&&(l[""]="http://www.w3.org/1999/xhtml"),l.xml=l.xml||"http://www.w3.org/XML/1998/namespace",e&&"string"==typeof e?o.parse(e,l,p):o.errorHandler.error("invalid doc source"),i.doc},r.prototype={startDocument:function(){this.doc=(new c).createDocument(null,null,null),this.locator&&(this.doc.documentURI=this.locator.systemId)},startElement:function(e,t,n,r){var a=this.doc,i=a.createElementNS(e,n||t),s=r.length;u(this,i),this.currentElement=i,this.locator&&o(this.locator,i);for(var c=0;c<s;c++){e=r.getURI(c);var l=r.getValue(c),f=(n=r.getQName(c),a.createAttributeNS(e,n));this.locator&&o(r.getLocator(c),f),f.value=f.nodeValue=l,i.setAttributeNode(f)}},endElement:function(e,t,n){var r=this.currentElement;r.tagName,this.currentElement=r.parentNode},startPrefixMapping:function(e,t){},endPrefixMapping:function(e){},processingInstruction:function(e,t){var n=this.doc.createProcessingInstruction(e,t);this.locator&&o(this.locator,n),u(this,n)},ignorableWhitespace:function(e,t,n){},characters:function(e,t,n){if(e=i.apply(this,arguments)){if(this.cdata)var r=this.doc.createCDATASection(e);else r=this.doc.createTextNode(e);this.currentElement?this.currentElement.appendChild(r):/^\s*$/.test(e)&&this.doc.appendChild(r),this.locator&&o(this.locator,r)}},skippedEntity:function(e){},endDocument:function(){this.doc.normalize()},setDocumentLocator:function(e){(this.locator=e)&&(e.lineNumber=0)},comment:function(e,t,n){e=i.apply(this,arguments);var r=this.doc.createComment(e);this.locator&&o(this.locator,r),u(this,r)},startCDATA:function(){this.cdata=!0},endCDATA:function(){this.cdata=!1},startDTD:function(e,t,n){var r=this.doc.implementation;if(r&&r.createDocumentType){var a=r.createDocumentType(e,t,n);this.locator&&o(this.locator,a),u(this,a)}},warning:function(e){console.warn("[xmldom warning]\t"+e,a(this.locator))},error:function(e){console.error("[xmldom error]\t"+e,a(this.locator))},fatalError:function(e){throw console.error("[xmldom fatalError]\t"+e,a(this.locator)),e}},"endDTD,startEntity,endEntity,attributeDecl,elementDecl,externalEntityDecl,internalEntityDecl,resolveEntity,getExternalSubset,notationDecl,unparsedEntityDecl".replace(/\w+/g,(function(e){r.prototype[e]=function(){return null}}));var s=Sn.XMLReader,c=t.DOMImplementation=Dr.DOMImplementation;t.XMLSerializer=Dr.XMLSerializer,t.DOMParser=n}));Ar.DOMImplementation,Ar.XMLSerializer;var Ir=Ar.DOMParser,Rr=function(){for(var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},t=e.length,n=void 0===t?10:t,r="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz",o="",a=0;a<n;a++)o+=r.charAt(Math.floor(Math.random()*r.length));return o},Cr=function(e){return(new Ir).parseFromString(e,"text/xml")},kr=function(){function e(t){c(this,e),p(this,"baseURL",void 0),this.baseURL=t}return f(e,[{key:"create",value:function(e,t,r){var o=arguments.length>3&&void 0!==arguments[3]?arguments[3]:{},a=o.permissions,i=void 0===a?"31":a,s=o.type,c=void 0===s?"0":s,l=u.default.request("POST","".concat(this.baseURL,"/ocs/v1.php/apps/files_sharing/api/v1/shares"),{shareType:c,shareWith:r,path:e,permissions:i},sn({headers:{"OCS-APIRequest":"true"}},{credential:t}));return n.check(l,{"share create":function(e){return 200===e.status}}),{id:Cr(l.body).getElementsByTagName("id")[0].childNodes[0].textContent,response:l}}},{key:"accept",value:function(e,t){var r=u.default.request("POST","".concat(this.baseURL,"/ocs/v1.php/apps/files_sharing/api/v1/shares/pending/").concat(e),void 0,sn({headers:{"OCS-APIRequest":"true"}},{credential:t}));return n.check(r,{"share accept":function(e){return 200===e.status}}),{response:r}}}]),e}(),jr="openIDConnect",Ur="basicAuth",Pr=new WeakMap,Lr=function(){function e(t){c(this,e),Pr.set(this,{writable:!0,value:void 0}),N(this,Pr,t)}return f(e,[{key:"credential",get:function(){return w(this,Pr)}}]),e}(),Mr=/\.|\[(?:[^[\]]*|(["'])(?:(?!\1)[^\\]|\\.)*?\1)\]/,qr=/^\w*$/;var Fr=function(e,t){if(bt(e))return!1;var n=typeof e;return!("number"!=n&&"symbol"!=n&&"boolean"!=n&&null!=e&&!$(e))||(qr.test(e)||!Mr.test(e)||null!=t&&e in Object(t))};function zr(e,t){if("function"!=typeof e||null!=t&&"function"!=typeof t)throw new TypeError("Expected a function");var n=function(){var r=arguments,o=t?t.apply(this,r):r[0],a=n.cache;if(a.has(o))return a.get(o);var i=e.apply(this,r);return n.cache=a.set(o,i)||a,i};return n.cache=new(zr.Cache||Ye),n}zr.Cache=Ye;var $r=zr;var Br=/[^.[\]]+|\[(?:(-?\d+(?:\.\d+)?)|(["'])((?:(?!\2)[^\\]|\\.)*?)\2)\]|(?=(?:\.|\[\])(?:\.|\[\]|$))/g,Vr=/\\(\\)?/g,Wr=function(e){var t=$r(e,(function(e){return 500===n.size&&n.clear(),e})),n=t.cache;return t}((function(e){var t=[];return 46===e.charCodeAt(0)&&t.push(""),e.replace(Br,(function(e,n,r,o){t.push(r?o.replace(Vr,"$1"):n||e)})),t}));var Xr=function(e,t){for(var n=-1,r=null==e?0:e.length,o=Array(r);++n<r;)o[n]=t(e[n],n,e);return o},Hr=R?R.prototype:void 0,Yr=Hr?Hr.toString:void 0;var Gr=function e(t){if("string"==typeof t)return t;if(bt(t))return Xr(t,e)+"";if($(t))return Yr?Yr.call(t):"";var n=t+"";return"0"==n&&1/t==-1/0?"-0":n};var Kr=function(e){return null==e?"":Gr(e)};var Zr=function(e,t){return bt(e)?e:Fr(e,t)?[e]:Wr(Kr(e))};var Qr=function(e){if("string"==typeof e||$(e))return e;var t=e+"";return"0"==t&&1/e==-1/0?"-0":t};var Jr=function(e,t){for(var n=0,r=(t=Zr(t,e)).length;null!=e&&n<r;)e=e[Qr(t[n++])];return n&&n==r?e:void 0};var eo=function(e,t,n){var r=null==e?void 0:Jr(e,t);return void 0===r?n:r},to=new WeakMap,no=new WeakMap,ro=new WeakMap,oo=new WeakMap,ao=new WeakMap,io=new WeakMap,uo=function(){function e(t,n){c(this,e),to.set(this,{writable:!0,value:void 0}),no.set(this,{writable:!0,value:void 0}),ro.set(this,{writable:!0,value:void 0}),oo.set(this,{writable:!0,value:void 0}),ao.set(this,{writable:!0,value:void 0}),io.set(this,{writable:!0,value:void 0}),N(this,to,t),N(this,no,n),N(this,ro,"".concat(n,"/oidc-callback.html")),N(this,oo,"".concat(n,"/signin/v1/identifier/_/logon")),N(this,ao,"".concat(n,"/konnect/v1/token"))}return f(e,[{key:"getContinueURI",value:function(){var e=u.default.post(w(this,oo),JSON.stringify({params:[w(this,to).login,w(this,to).password,"1"],hello:{scope:"openid profile email",client_id:"web",redirect_uri:w(this,ro),flow:"oidc"},state:"22db4af49efce268"}),{headers:{"Kopano-Konnect-XSRF":"1",Referer:w(this,no),"Content-Type":"application/json"}});console.log("###"),console.log(e),console.log("###");var t=eo(e.json(),"hello.continue_uri");return 200===e.status&&t||n.fail(w(this,oo)),t}},{key:"getCode",value:function(e){var t,r="".concat(e,"?").concat((t={client_id:"web",prompt:"none",redirect_uri:w(this,ro),response_mode:"query",response_type:"code",scope:"openid profile email"},Object.keys(t).map((function(e){return encodeURIComponent(e)+"="+encodeURIComponent(t[e])})).join("&"))),o=u.default.get(r,{redirects:0}),a=eo(function(e){try{e=new URL(e).search}catch(e){}return Object.fromEntries(new URLSearchParams(e))}(o.headers.Location),"code");return 302===o.status&&a||n.fail(e),a}},{key:"getToken",value:function(e){var t=u.default.post(w(this,ao),{client_id:"web",code:e,redirect_uri:w(this,ro),grant_type:"authorization_code"}),r={accessToken:eo(t.json(),"access_token"),tokenType:eo(t.json(),"token_type"),idToken:eo(t.json(),"id_token"),expiresIn:eo(t.json(),"expires_in")};return 200===t.status&&r.accessToken&&r.tokenType&&r.idToken&&r.expiresIn||n.fail(w(this,ao)),r}},{key:"credential",get:function(){if(!w(this,io)||w(this,io).validTo<=new Date){var e=this.getContinueURI(),t=this.getCode(e),n=this.getToken(t);N(this,io,{validTo:(r=new Date,r.setSeconds(r.getSeconds()+n.expiresIn-5),r),token:n})}var r;return w(this,io).token}}]),e}(),so=new WeakMap,co=function(){function e(t,n,r){switch(c(this,e),so.set(this,{writable:!0,value:void 0}),n){case jr:N(this,so,new uo(t,r));break;case Ur:N(this,so,new Lr(t))}}return f(e,[{key:"credential",get:function(){return w(this,so).credential}}]),e}(),lo=function(){function e(t){c(this,e),p(this,"baseURL",void 0),this.baseURL=t}return f(e,[{key:"get",value:function(e,t){var n=new co({login:e.login,password:e.password},t,this.baseURL).credential;return{user:{login:e.login,password:e.password,credential:n}}}},{key:"enable",value:function(e,t){var r=u.default.request("PUT","".concat(this.baseURL,"/ocs/v1.php/cloud/users/").concat(e.login,"/enable"),void 0,sn({headers:{"OCS-APIRequest":"true"}},{credential:t}));return n.check(r,{"user enable":function(e){return 200===e.status}}),{response:r}}},{key:"create",value:function(e,t,r){var o=u.default.request("POST","".concat(this.baseURL,"/ocs/v1.php/cloud/users"),{userid:e.login,password:e.password,email:"".concat(e.login,"@owncloud.org")},sn({headers:{"OCS-APIRequest":"true"}},{credential:t}));return n.check(o,{"user create":function(e){return 200===e.status}}),{enableResponse:this.enable(e,t).response,user:this.get(e,r).user,createResponse:o}}},{key:"delete",value:function(e,t){var r=u.default.request("DELETE","".concat(this.baseURL,"/ocs/v1.php/cloud/users/").concat(e),void 0,sn({headers:{"OCS-APIRequest":"true"}},{credential:t}));return n.check(r,{"user delete":function(e){return 200===e.status}}),{response:r}}}]),e}(),fo=function(e){v(r,lo);var t=y(r);function r(){return c(this,r),t.apply(this,arguments)}return f(r,[{key:"create",value:function(e,t,r){var o=u.default.request("POST","".concat(this.baseURL,"/graph/v1.0/users"),JSON.stringify({onPremisesSamAccountName:e.login,displayName:e.login,mail:"".concat(e.login,"@owncloud.org"),passwordProfile:{password:e.password}}),sn({},{credential:t}));return n.check(o,{"user create":function(e){return 200===e.status}}),{user:this.get(e,r).user,createResponse:o}}},{key:"delete",value:function(e,t){var r=u.default.request("DELETE","".concat(this.baseURL,"/graph/v1.0/users/").concat(e),void 0,sn({},{credential:t}));return n.check(r,{"user delete":function(e){return 204===e.status}}),{response:r}}}]),r}(),po="latest",ho="legacy",vo={baseURL:__ENV.BASE_URL||"https://localhost:9200",authAdapter:__ENV.AUTH_ADAPTER==Ur?Ur:jr,apiVersion:__ENV.API_VERSION==ho?ho:po,testFolder:"test-folder",adminUser:{login:__ENV.ADMIN_LOGIN||"admin",password:__ENV.ADMIN_PASSWORD||"admin"},assets:{size:1024,quantity:10},k6:{vus:5,iterations:50,insecureSkipTLSVerify:!0,thresholds:{"http_req_duration{scenario:default}":["p(99)<300","p(99.9)<500","max<1000"]}}},mo=new function e(t,n){switch(c(this,e),p(this,"dav",void 0),p(this,"share",void 0),p(this,"user",void 0),p(this,"me",void 0),this.share=new kr(t),n){case po:this.dav=new ln(t),this.user=new fo(t),this.me=new pn(t);break;case ho:this.dav=new cn(t),this.user=new lo(t),this.me=new fn(t)}}(vo.baseURL,vo.apiVersion),go={vus:5,iterations:50,insecureSkipTLSVerify:!0};exports.default=function(t){var n=t.userInfos[i.default.vu.idInTest-1],r=mo.user.get(n,vo.authAdapter).user,o=[i.default.scenario.iterationInTest,"initial",r.login].join("-");mo.dav.create(n.home,o,r.credential);var a=e.randomBytes(1024*vo.assets.size);Q(vo.assets.quantity,(function(e){mo.dav.upload(n.home,[o,e].join("/"),a,r.credential)}));var u=[i.default.scenario.iterationInTest,"final",r.login].join("-");mo.dav.move(n.home,o,u,r.credential)},exports.options=go,exports.setup=function(){var e=mo.user.get({login:vo.adminUser.login,password:vo.adminUser.password},vo.authAdapter).user,t=h(h({},e),{},{home:String(mo.me.driveInfo(e.login,e.credential,{selector:'value.#(driveType=="personal").id'}))});return mo.dav.create(t.home,vo.testFolder,e.credential),{adminInfo:t,userInfos:Q(go.vus||1,(function(){var t=mo.user.create({login:Rr(),password:Rr()},e.credential,vo.authAdapter).user,n=mo.share.create(vo.testFolder,e.credential,t.login);return mo.share.accept(n.id,t.credential),h(h({},t),{},{home:String(mo.me.driveInfo([t.login,vo.testFolder].join("/"),t.credential,{selector:'value.#(driveAlias=="mountpoint/'.concat(vo.testFolder,'").id')}))})}))}},exports.teardown=function(e){var t=e.adminInfo,n=e.userInfos,r=mo.user.get(t,vo.authAdapter).user;mo.dav.delete(t.home,vo.testFolder,r.credential),n.forEach((function(e){return mo.user.delete(e.login,r.credential)}))};
