+++ src/lib/api/share.ts
@@ -1,49 +1,60 @@
-import { RefinedResponse, ResponseType } from 'k6/http';
+import { check } from 'k6';
+import http from 'k6/http';
 
-import * as types from '../types';
-import * as api from './api';
+import { Credential } from '../auth';
+import { parseXML } from '../utils';
+import { Result } from './api';
+import { buildParams } from './utils';
 
-export class Create {
-    public static exec({
-        credential,
-        shareType,
-        shareWith,
+export class ShareAPI {
+  protected baseURL: string;
+
+  constructor(baseURL: string) {
+    this.baseURL = baseURL;
+  }
+
+  create(
+    path: string,
+    credential: Credential,
+    recipient: string,
+    { permissions = '31', type = '0' }: { permissions?: string; type?: string } = {},
+  ): { id: string } & Result {
+    const createResponse = http.request(
+      'POST',
+      `${this.baseURL}/ocs/v1.php/apps/files_sharing/api/v1/shares`,
+      {
+        shareType: type,
+        shareWith: recipient,
         path,
         permissions,
-        tags,
-    }: {
-        credential: types.Credential;
-        shareType: string;
-        shareWith: string;
-        path: string;
-        permissions: string;
-        tags?: types.Tags;
-    }): RefinedResponse<ResponseType> {
-        return api.request({
-            method: 'POST',
-            credential,
-            path: `/ocs/v1.php/apps/files_sharing/api/v1/shares`,
-            params: { tags },
-            body: { shareType, shareWith, path, permissions },
-        });
-    }
-}
+      },
+      buildParams({ headers: { 'OCS-APIRequest': 'true' } }, { credential }),
+    );
 
-export class Accept {
-    public static exec({
-        credential,
-        id,
-        tags,
-    }: {
-        credential: types.Credential;
-        id: string;
-        tags?: types.Tags;
-    }): RefinedResponse<ResponseType> {
-        return api.request({
-            method: 'POST',
-            credential,
-            path: `/ocs/v1.php/apps/files_sharing/api/v1/shares/pending/${id}`,
-            params: { tags },
-        });
-    }
+    check(createResponse, {
+      'share create': ({ status }) => status === 200,
+    });
+
+    return {
+      id: parseXML(createResponse.body).getElementsByTagName('id')[0].childNodes[0].textContent!,
+      response: createResponse,
+    };
+  }
+
+  accept(id: string, credential: Credential): Result {
+    const acceptResponse = http.request(
+      'POST',
+      `${this.baseURL}/ocs/v1.php/apps/files_sharing/api/v1/shares/pending/${id}`,
+      undefined,
+      buildParams({ headers: { 'OCS-APIRequest': 'true' } }, { credential }),
+    );
+
+    check(acceptResponse, {
+      'share accept': ({ status }) => status === 200,
+    });
+
+    return {
+      response: acceptResponse,
+    };
+  }
 }
