--- src/lib/index.ts
@@ -1,9 +1,0 @@
-import 'core-js';
-
-export * as api from './api';
-export { default as auth } from './auth';
-export * as defaults from './defaults';
-export * as k6 from './k6';
-export * as playbook from './playbook';
-export * as types from './types';
-export * as utils from './utils';
