--- src/lib/k6/index.ts
@@ -1,1 +1,0 @@
-export { default as options } from './options';
