--- src/lib/playbook/index.ts
@@ -1,3 +1,0 @@
-export * as dav from './dav';
-export * as share from './share';
-export * as users from './users';
